Claimants are shown a timeout warning dialog after inactivity
Pull in an external dependency for showing the timeout dialog. The
library `a11y-dialog` lets us attach to dialog events.

I think it makes sense to include the timeout dialog on every page,
incase the user finds themselves on a non-claim page their session will
still timeout after 30 minutes so we still need display the dialog if
they have a claim in-progress.

I've attached the refresh session AJAX call to the dialog hide event so
that if a user has closed the dialog through other means
(pressing escape or close) the session is still refreshed as they are
still active.