Add headless chrome for testing JavaScript features
OmniAuth needs to be told to `redirect_to_failure` in the specs
otherwise it will raise an exception.

We need to whitelist the chromedriver download URL so that the
`webdrivers` gem can download the executable.

Installed `launchy` gem so that we can use the Capybara helper
`save_and_open_screenshot`