Refactor the ineligble screen to use partials
We need more flexibility on the ineligible screens as we would like to
use html and not just a single string when defining the ineligible
message.

The TSLR ineligble messages have now been changed over to use partial
views and a new ineligble message created for if you don't teach an
eligible subject.