Create static pages for Verify return states
Created a series of pages to handle the various Verify return states and
try and provide a way for users to recover from errors. This should be
in addition to the information that Verify hub will show users to help
them recover.

Views created should handle the 3 error states and the success state, as
we will probably want to play back the users info Verify provided so
when they get to the check answers page they aren't suprised by it. This
raises the issue what if they want to change some information but we
will have to think about that at a later date.